Removed unnecessary spacing and modified colors
Unnecessary spacing was removed to make the `starship` prompt more
compact while preserving

The `starship` prompt has been modified to be more compact while
preserving and, in some cases, improving the legibility of various
items.

The prompt works well enough as-is, but it would be preferable to
customize the format such that the `username`@`hostname` "group" is
displayed before the prompt character(s) on the second line.